move gem.config.Arbitraries into gem.arb.ArbStaticConfig and gem.arb.ArbDynamicConfig
